move file to another directory
