+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:invenza/pages/home_page.dart';
-import 'package:invenza/pages/login_page.dart';
+import 'package:invenza/pages/login_page/login_page.dart';
 import 'package:invenza/pages/routes.dart';
 import 'package:invenza/providers/auth_provider.dart';
 import 'package:invenza/theme/theme.dart';
