+++ lib/pages/routes.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:invenza/pages/home_page.dart';
-import 'package:invenza/pages/login_page.dart';
+import 'package:invenza/pages/login_page/login_page.dart';
 
 final Map<String, WidgetBuilder> routes = {
   '/login' : (context) => const LoginPage(),
