+++ lib/providers/api_provider.dart
@@ -4,6 +4,6 @@ import '../services/api_client.dart';
 import 'log_provider.dart';
 
 final apiClientProvider = Provider<ApiClient>((ref) {
-  final logger = ref.read(logProvider); // ✅ 依賴注入
+  final logger = ref.read(logProvider);
   return ApiClient(logger);
 });
